API security plus UI updates

+++ 4thYearProject.Api/Controllers/FollowingController.cs
@@ -1,4 +1,8 @@
-﻿namespace _4thYearProject.Api.Controllers
+﻿using System.Linq;
+using System.Threading.Tasks;
+using _4thYearProject.Shared;
+
+namespace _4thYearProject.Api.Controllers
 {
     using _4thYearProject.Api.Models;
     using _4thYearProject.Shared.Models;
@@ -12,15 +16,17 @@
         private readonly IFollowingRepository _followingRepository;
 
         private readonly IWebHostEnvironment env;
+        private readonly IUserService _userService;
 
-        public FollowingController(IFollowingRepository followingRepository, IWebHostEnvironment env)
+        public FollowingController(IFollowingRepository followingRepository, IWebHostEnvironment env, IUserService userService)
         {
             _followingRepository = followingRepository;
             this.env = env;
+            _userService = userService;
         }
 
         [HttpPost]
-        public IActionResult AddFollowing([FromBody] Following following)
+        public async Task<IActionResult> AddFollowing([FromBody] Following following)
         {
             if (following == null)
                 return BadRequest();
@@ -28,34 +34,46 @@
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
+            var identity = await _userService.GetUserAsync();
+
+            if (identity == null)
+                return Unauthorized();
+
+            string LoggedInID = identity.Claims.Where(c => c.Type.Equals("sub"))
+                .Select(c => c.Value).SingleOrDefault().ToString();
+
+            if (LoggedInID != following.Follower_ID)
+                return Unauthorized();
+
             var createdFollowing = _followingRepository.AddFollowing(following);
 
             return
                 Created("following",
-                    createdFollowing); // I don't think this is necessary. Look up the one for a button.
+                    createdFollowing);
         }
 
         [HttpDelete("{Follower_ID}/{Followed_ID}")]
-        public IActionResult RemoveFollowing(string Follower_ID, string Followed_ID)
+        public async Task<IActionResult> RemoveFollowing(string Follower_ID, string Followed_ID)
         {
             if ((Follower_ID == string.Empty) ^ (Followed_ID == string.Empty))
                 return BadRequest();
+
+            var identity = await _userService.GetUserAsync();
+
+            if (identity == null)
+                return Unauthorized();
+
+            string LoggedInID = identity.Claims.Where(c => c.Type.Equals("sub"))
+                .Select(c => c.Value).SingleOrDefault().ToString();
+
+            if (LoggedInID != Follower_ID)
+                return Unauthorized();
 
             _followingRepository.RemoveFollowing(Follower_ID, Followed_ID);
 
             return NoContent(); //success
         }
 
-        //[HttpDelete("{Follower_ID}/{Followed_ID}")]
-        //public IActionResult RemoveFollowing(string Follower_ID, string Followed_ID)
-        //{
-        //    if (Follower_ID == string.Empty ^ Followed_ID == string.Empty)
-        //        return BadRequest();
-
-        //    _followingRepository.RemoveFollowing(Follower_ID, Followed_ID);
-
-        //    return NoContent();//success
-        //}
         [HttpGet("{Follower_ID}/{Followed_ID}")]
         public IActionResult VerifyFollowing(string Follower_ID, string Followed_ID)
         {
